Support Dumping From Disk & Code Cleanup
Dumper can now dump directly from disk (dumped binaries/modules). Drag & Drop is supported for multiple config files now.
