+++ GH-Offset-Dumper.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.35931.194
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "GH-Offset-Dumper", "GH-Offset-Dumper.vcxproj", "{1722F02C-911E-44E4-8613-37AB98D08C22}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{1722F02C-911E-44E4-8613-37AB98D08C22}.Debug|x64.ActiveCfg = Debug|x64
+		{1722F02C-911E-44E4-8613-37AB98D08C22}.Debug|x64.Build.0 = Debug|x64
+		{1722F02C-911E-44E4-8613-37AB98D08C22}.Debug|x86.ActiveCfg = Debug|Win32
+		{1722F02C-911E-44E4-8613-37AB98D08C22}.Debug|x86.Build.0 = Debug|Win32
+		{1722F02C-911E-44E4-8613-37AB98D08C22}.Release|x64.ActiveCfg = Release|x64
+		{1722F02C-911E-44E4-8613-37AB98D08C22}.Release|x64.Build.0 = Release|x64
+		{1722F02C-911E-44E4-8613-37AB98D08C22}.Release|x86.ActiveCfg = Release|Win32
+		{1722F02C-911E-44E4-8613-37AB98D08C22}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {113DE095-A7DD-4C4E-A0C5-0AEB79D6BB8D}
+	EndGlobalSection
+EndGlobal
