+++ GH-Offset-Dumper.vcxproj
@@ -75,24 +75,32 @@
     <GenerateManifest>true</GenerateManifest>
     <IncludePath>$(DXSDK_DIR)Include;$(IncludePath)</IncludePath>
     <LibraryPath>$(DXSDK_DIR)Lib\x64;$(DXSDK_DIR)Lib\x86;$(LibraryPath)</LibraryPath>
+    <OutDir>$(SolutionDir)bin</OutDir>
+    <TargetName>$(ProjectName)-x$(PlatformArchitecture)</TargetName>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
     <LinkIncremental>true</LinkIncremental>
     <GenerateManifest>true</GenerateManifest>
     <IncludePath>$(DXSDK_DIR)Include;$(IncludePath)</IncludePath>
     <LibraryPath>$(DXSDK_DIR)Lib\x64;$(DXSDK_DIR)Lib\x86;$(LibraryPath)</LibraryPath>
+    <OutDir>$(SolutionDir)bin</OutDir>
+    <TargetName>$(ProjectName)-x$(PlatformArchitecture)</TargetName>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
     <LinkIncremental>false</LinkIncremental>
     <GenerateManifest>true</GenerateManifest>
     <IncludePath>$(DXSDK_DIR)Include;$(IncludePath)</IncludePath>
     <LibraryPath>$(DXSDK_DIR)Lib\x64;$(DXSDK_DIR)Lib\x86;$(LibraryPath)</LibraryPath>
+    <TargetName>$(ProjectName)-x$(PlatformArchitecture)</TargetName>
+    <OutDir>$(SolutionDir)bin</OutDir>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
     <LinkIncremental>false</LinkIncremental>
     <GenerateManifest>true</GenerateManifest>
     <IncludePath>$(DXSDK_DIR)Include;$(IncludePath)</IncludePath>
     <LibraryPath>$(DXSDK_DIR)Lib\x64;$(DXSDK_DIR)Lib\x86;$(LibraryPath)</LibraryPath>
+    <OutDir>$(SolutionDir)bin</OutDir>
+    <TargetName>$(ProjectName)-x$(PlatformArchitecture)</TargetName>
   </PropertyGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
     <ClCompile>
@@ -163,24 +171,28 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="FileScanner.cpp" />
+    <ClCompile Include="GHDumper.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="zip.c" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="FileScanner.h" />
     <ClInclude Include="GHDumper.h" />
     <ClInclude Include="json.hpp" />
+    <ClInclude Include="resource.h" />
     <ClInclude Include="zip.h" />
     <ClInclude Include="miniz.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="config.json" />
-    <None Include="README.md" />
+    <None Include="diskSample.json" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="GH-Offset-Dumper.rc" />
   </ItemGroup>
   <ItemGroup>
-    <Image Include="icon-96px.ico" />
+    <Image Include="gh.ico" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
